+++ src/definitions/Hubot.d.ts
