+++ src/definitions/ICallback.d.ts
@@ -1,0 +1,6 @@
+import { IContext } from "./IContext";
+
+export interface ICallback {
+    (context: IContext): any;
+  }
+
