+++ src/definitions/index.ts
@@ -1,0 +1,20 @@
+import { IParameter } from "./IParameter";
+import { ICallback } from "./ICallback";
+import { ICommand } from "./ICommand";
+import { IHelpCommand } from "./IHelpCommand";
+import { IMutable } from "./IMutable";
+import { IContext } from "./IContext";
+import { IParameterValueCollection } from "./IParameterValueCollection";
+import { ITool } from "./ITool";
+
+
+export {
+    ICallback,
+    ICommand,
+    IContext,
+    IHelpCommand,
+    IMutable,
+    IParameter,
+    IParameterValueCollection,
+    ITool
+}
