+++ src/entities/parameters/FractionParameter.ts
