+++ src/entities/parameters/FractionStyle.ts
