+++ src/entities/parameters/IPv4Parameter.ts
