+++ src/entities/parameters/NumberStyle.ts
