+++ src/entities/parameters/RegExStringParameter.ts
