+++ src/entities/parameters/StringParameter.ts
